backend/pyproject.toml
[project]
name = "test-django-react-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-allauth>=65.7.0",
    "django-environ>=0.12.0",
    "django-guardian>=2.4.0",
    "django-vite>=3.1.0",
    "inertia-django>=1.2.0",
    "pydantic>=2.11.2",
    "rules>=3.5",
]

[tool.ruff]
target-version = "py312"
exclude = [
    ".venv",
    "migrations",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle (error)
    "F", # pyflakes
    "B", # bugbear
    "B9",
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
    "I", # isort
    "UP", # pyupgrade
    "PIE", # flake8-pie
    "PGH", # pygrep-hooks
    "PYI", # flake8-pyi
    "RUF",
]

[tool.ruff.format]
